De prijs in 'int' neergezet in plaats van 'float', omdat we bij het overzetten de prijs in centen laten

+++ ddl.sql
@@ -16,7 +16,7 @@ CREATE TABLE Categories (categoryid SERIAL, category varchar(255) NOT NULL, PRIM
 CREATE TABLE Events (eventid SERIAL, Sessionssessionid varchar(255) NOT NULL, Productsproductid varchar(255), PRIMARY KEY (eventid));
 CREATE TABLE Genders (genderid SERIAL, gender varchar(255), PRIMARY KEY (genderid));
 CREATE TABLE Orderedproducts (orderedproductid SERIAL, Sessionssessionid varchar(255) NOT NULL, Productsproductid varchar(255) NOT NULL, PRIMARY KEY (orderedproductid));
-CREATE TABLE Products (productid varchar(255) NOT NULL, name varchar(255), description varchar(1023), price float, herhaalaankopen BIT, recommendable BIT, Brandsbrandid int, Categoriescategoryid int, Sub_sub_categoriessub_sub_categoryid int, Sub_categoriessub_categoryid int, Gendersgenderid int, PRIMARY KEY (productid));
+CREATE TABLE Products (productid varchar(255) NOT NULL, name varchar(255), description varchar(1023), price int, herhaalaankopen BIT, recommendable BIT, Brandsbrandid int, Categoriescategoryid int, Sub_sub_categoriessub_sub_categoryid int, Sub_categoriessub_categoryid int, Gendersgenderid int, PRIMARY KEY (productid));
 CREATE TABLE Profiles (profileid varchar(255) NOT NULL, PRIMARY KEY (profileid));
 CREATE TABLE Recommendedproducts (recommendedproductid SERIAL, Profilesprofileid varchar(255) NOT NULL, Productsproductid varchar(255) NOT NULL, PRIMARY KEY (recommendedproductid));
 CREATE TABLE Sessions (sessionid varchar(255) NOT NULL, Profilesprofileid varchar(255), sessionstart timestamp NOT NULL, sessionend timestamp NOT NULL, has_sale BIT NOT NULL, PRIMARY KEY (sessionid));
